server/addons/product_ai_search/requirements.txt
# Dependências do Product AI Search Module para Odoo 17
# Para instalar: pip install -r requirements.txt

# LlamaIndex - Framework principal
llama-index>=0.10.0
llama-index-vector-stores-elasticsearch>=0.1.0
llama-index-embeddings-openai>=0.1.0

# Vector Stores
elasticsearch>=8.0.0,<9.0.0  # Para Elasticsearch

# OpenAI
openai>=1.0.0

# Processamento de imagens
Pillow>=9.0.0
requests>=2.28.0

# Utilitários
pandas>=1.5.0
tqdm>=4.64.0
python-dotenv>=1.0.0

# Dependências opcionais para melhor performance
numpy>=1.24.0
scipy>=1.10.0
